openai_tutorial/requirements.txt
streamlit
openai
langchain
langchain-experimental
langchain-openai
langchain-core
langchain-community
pypdf
pymupdf
pdfplumber
chromadb
faiss-cpu
tiktoken
python-dotenv
googletrans
pdf2image
pytesseract
Pillow
python-docx
jupyterlab
sounddevice
scipy
numpy
wavio
pyaudio
pygame
pandas
matplotlib
seaborn
gradio
pytubefix